changed discount value type from decimal to int discount patch works

+++ database/migrations/2021_04_13_140020_create_discounts_table.php
@@ -16,7 +16,7 @@ class CreateDiscountsTable extends Migration
         Schema::create('discounts', function (Blueprint $table) {
             $table->id();
             $table->string('code')->unique();
-            $table->decimal('value');
+            $table->integer('value');
             $table->foreignId('user_id')->constrained();
             $table->softDeletes();
             $table->timestamps();
